version 0.5.0 alpha
- refactor
- use git revision count as build number for app building
- use git tag for version string in app building
