+++ src/configs/default_env.js
@@ -1,0 +1,9 @@
+
+export default ({ config }) => ({
+  PORT: 8080,
+  MONGO_URL: `mongodb://localhost/${config.appname}`,
+  MONGO_OPLOG_URL: 'mongodb://localhost/local',
+  MAIL_URL: 'smtp://localhost:25',
+  METEOR_SETTINGS: {
+  },
+});
