--- src/logs.js
@@ -1,8 +1,0 @@
-import colors from 'colors/safe';
-
-export const intro = line => console.log(colors.yellow(line));
-export const actionTitle = (title) => {
-  intro('');
-  intro(`🐱 🔧 ${title}`);
-  intro('');
-};
