+++ src/ui/action_title.js
@@ -1,0 +1,7 @@
+import printYellow from './print_yellow';
+
+export default (title) => {
+  printYellow('');
+  printYellow(`🐱 🔧 ${title}`);
+  printYellow('');
+};
