+++ src/ui/done_error.js
@@ -1,0 +1,13 @@
+import printYellow from './print_yellow';
+
+export default (error, message) => {
+  printYellow('');
+  printYellow('');
+  printYellow(`╗ 🙀  ${message}  😿`);
+  printYellow('╚══════════════════════════════════════════════════');
+  printYellow('😾         🐁 🐁 🐁 🐁 🐁 🐁');
+  console.log(`${error && (error.message || error.reason)}`);
+  printYellow('');
+  console.log(error && error.stack);
+  printYellow('═══════════════════════════════════════════════════');
+};
