+++ src/ui/done_success.js
@@ -1,0 +1,9 @@
+import printYellow from './print_yellow';
+
+export default (message) => {
+  printYellow('');
+  printYellow('');
+  printYellow('╗');
+  printYellow(`╚═╗ ${message}  👋 🐱`);
+  printYellow('  ╚════════════════════════════════════════════════');
+};
