+++ src/ui/print_yellow.js
@@ -1,0 +1,3 @@
+import colors from 'colors/safe';
+
+export default line => console.log(colors.yellow(line));
